router/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
psycopg2-binary
python-jose
pandas
meteostat
holidays
python-multipart
